Split up test/integration/index.ts
