+++ bokehjs/test/integration/regressions.ts
@@ -1,0 +1,150 @@
+import {display, fig, row} from "./utils"
+
+import {
+  Arrow, ArrowHead, NormalHead, BoxAnnotation,
+  DataRange1d, FactorRange,
+  ColumnDataSource,
+} from "@bokehjs/models"
+
+import {Color} from "@bokehjs/core/types"
+import {Anchor} from "@bokehjs/core/enums"
+
+
+describe("Bug", () => {
+  describe("in issue #9879", () => {
+    it("disallows to change FactorRange to a lower dimension with a different number of factors", async () => {
+      const p = fig([200, 200], {
+        title: null,
+        toolbar_location: null,
+        x_range: new FactorRange({factors: [["a", "b"], ["b", "c"]]}),
+        y_range: new DataRange1d(),
+      })
+      const source = new ColumnDataSource({data: {x: [["a", "b"], ["b", "c"]], y: [1, 2]}})
+      p.vbar({x: {field: "x"}, top: {field: "y"}, source})
+      const view = await display(p, [250, 250])
+
+      source.data = {x: ["a"], y: [1]}
+      ;(p.x_range as FactorRange).factors = ["a"]
+      await view.ready
+    })
+  })
+
+  describe("in issue #9522", () => {
+    it("disallows arrow to be positioned correctly in stacked layouts", async () => {
+      const horz = (end?: ArrowHead) => new Arrow({x_start: 1, x_end: 5, y_start: 0, y_end:  0, end})
+      const vert = (end?: ArrowHead) => new Arrow({x_start: 2, x_end: 2, y_start: 1, y_end: -2, end})
+
+      const p1 = fig([200, 200], {x_range: [0, 6], y_range: [-3, 2]})
+      p1.add_layout(horz(new NormalHead({fill_color: "blue"})))
+      p1.add_layout(vert())
+
+      const p2 = fig([200, 200], {x_range: [0, 6], y_range: [-3, 2]})
+      p2.add_layout(horz())
+      p2.add_layout(vert(new NormalHead({fill_color: "green"})))
+
+      await display(row([p1, p2]), [450, 250])
+    })
+  })
+
+  describe("in issue #9703", () => {
+    it("disallows ImageURL glyph to set anchor and angle at the same time", async () => {
+      const p = fig([300, 300], {x_range: [-1, 10], y_range: [-1, 10]})
+
+      const svg = `\
+<svg version="1.1" viewBox="0 0 2 2" xmlns="http://www.w3.org/2000/svg">
+  <path d="M 0,0 2,0 1,2 Z" fill="green" />
+</svg>
+`
+      const img = `data:image/svg+xml;utf-8,${svg}`
+
+      let y = 0
+      const w = 1, h = 1
+
+      for (const anchor of Anchor) {
+        p.image_url({url: [img], x: 0, y, w, h, anchor, angle: 0})
+        p.image_url({url: [img], x: 1, y, w, h, anchor, angle: Math.PI/6})
+        p.image_url({url: [img], x: 2, y, w, h, anchor, angle: Math.PI/4})
+        p.image_url({url: [img], x: 3, y, w, h, anchor, angle: Math.PI/3})
+        p.image_url({url: [img], x: 4, y, w, h, anchor, angle: Math.PI/2})
+        p.image_url({url: [img], x: 5, y, w, h, anchor, angle: Math.PI/1})
+        y += 1
+      }
+
+      await display(p, [350, 350])
+    })
+  })
+
+  describe("in issue #9724", () => {
+    it("makes automatic padding in data ranges inconsistent", async () => {
+      const x = [0.1, 0.5, 1.0, 1.5, 2.0, 2.5, 3.0]
+
+      const padding = {
+        range_padding: 1,
+        range_padding_units: "absolute" as const,
+      }
+
+      const p0 = (() => {
+        const x_range = new DataRange1d()
+        const y_range = new DataRange1d(padding)
+        const p = fig([150, 150], {x_range, y_range})
+        p.line({x, y: 10, line_width: 2, color: "red"})
+        return p
+      })()
+
+      const p1 = (() => {
+        const x_range = new DataRange1d()
+        const y_range = new DataRange1d(padding)
+        const p = fig([150, 150], {x_range, y_range})
+        p.line({x, y: 10, line_width: 2, color: "red"})
+        p.line({x, y: 15, line_width: 2, color: "blue"})
+        return p
+      })()
+
+      const p2 = (() => {
+        const x_range = new DataRange1d()
+        const y_range = new DataRange1d({start: 0, ...padding})
+        const p = fig([150, 150], {x_range, y_range})
+        p.line({x, y: 10, line_width: 2, color: "red"})
+        return p
+      })()
+
+      const p3 = (() => {
+        const x_range = new DataRange1d()
+        const y_range = new DataRange1d({start: 0, ...padding})
+        const p = fig([150, 150], {x_range, y_range})
+        p.line({x, y: 10, line_width: 2, color: "red"})
+        p.line({x, y: 15, line_width: 2, color: "blue"})
+        return p
+      })()
+
+      await display(row([p0, p1, p2, p3]), [650, 200])
+    })
+  })
+
+  describe("in issue #9877", () => {
+    function plot(fill: Color | null, line: Color | null) {
+      const p = fig([200, 200], {x_range: [0, 3], y_range: [0, 3]})
+      p.circle({x: [1, 1, 2, 2], y: [1, 2, 1, 2], radius: 0.5, line_color: null, fill_color: "red"})
+
+      const box = new BoxAnnotation({
+        bottom: 1, top: 2, left: 1, right: 2,
+        fill_color: fill, fill_alpha: 0.5,
+        line_color: line, line_alpha: 1.0, line_width: 4,
+      })
+      p.add_layout(box)
+      return p
+    }
+
+    it("disallows BoxAnnotation to respect fill_color == null", async () => {
+      const p0 = plot("blue", "green")
+      const p1 = plot(null, "green")
+      await display(row([p0, p1]), [450, 250])
+    })
+
+    it("disallows BoxAnnotation to respect line_color == null", async () => {
+      const p0 = plot("blue", "green")
+      const p1 = plot("blue", null)
+      await display(row([p0, p1]), [450, 250])
+    })
+  })
+})
